requirements.txt
python-telegram-bot
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
google-generativeai
python-dotenv
pydantic
pydantic-settings
dateparser
aiohttp
tenacity
pyasn1<0.5.0,>=0.4.1
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0

